refactor: move resolve fitler to more local scope
